Refactor and better alignment of radio button focus
