+++ components/shared/components/Widget/components/panes/DonationPane/ShareSelector/Multiple/MultipleCauseAreasSelector.tsx
@@ -1,24 +1,22 @@
 import { useDispatch, useSelector } from "react-redux";
-import { State } from "../../../../store/state";
+import { State } from "../../../../../store/state";
 import {
   CauseAreaShareSelectionTitle,
   CauseAreaShareSelectionTitleSmartDistributionWrapper,
   CauseAreaShareSelectionTitleWrapper,
   PercentageInputWrapper,
 } from "./MultipleCauseAreasSelector.style";
-import { Toggle } from "../../../shared/Toggle/Toggle";
-import { setShareType } from "../../../../store/donation/actions";
-import { ShareType } from "../../../../types/Enums";
+import { Toggle } from "../../../../shared/Toggle/Toggle";
+import { setShareType } from "../../../../../store/donation/actions";
+import { ShareType } from "../../../../../types/Enums";
 import AnimateHeight from "react-animate-height";
-import { SharesSelection } from "./ShareSelection";
-import { SharesSum } from "./SharesSum";
+import { SharesSelection } from "../ShareSelection";
+import { SharesSum } from "../SharesSum";
 import { useRef } from "react";
-import { TextInput } from "../../../shared/Input/TextInput";
 
 export const MultipleCauseAreasSelector: React.FC = () => {
   const layout = useSelector((state: State) => state.layout);
   const donation = useSelector((state: State) => state.donation);
-  const containerRef = useRef<HTMLDivElement>(null);
   const dispatch = useDispatch();
 
   if (!layout.causeAreas) return null;
@@ -31,7 +29,7 @@ export const MultipleCauseAreasSelector: React.FC = () => {
         )?.shareType;
 
         return (
-          <div key={causeAreaOrgs.name} ref={containerRef}>
+          <div key={causeAreaOrgs.name}>
             <CauseAreaShareSelectionTitleWrapper>
               <CauseAreaShareSelectionTitle>{causeAreaOrgs.name}</CauseAreaShareSelectionTitle>
               <CauseAreaShareSelectionTitleSmartDistributionWrapper>
@@ -44,13 +42,6 @@ export const MultipleCauseAreasSelector: React.FC = () => {
                   onChange={(checked) => {
                     if (checked) {
                       dispatch(setShareType(causeAreaOrgs.name, ShareType.CUSTOM));
-                      // Smooth scroll to title when toggling to custom
-                      setTimeout(() => {
-                        containerRef.current?.scrollIntoView({
-                          behavior: "smooth",
-                          block: "center",
-                        });
-                      }, 300);
                     } else {
                       dispatch(setShareType(causeAreaOrgs.name, ShareType.STANDARD));
                     }
@@ -65,10 +56,7 @@ export const MultipleCauseAreasSelector: React.FC = () => {
               </span>
             </PercentageInputWrapper>
 
-            <AnimateHeight height={shareType == ShareType.CUSTOM ? "auto" : 0} animateOpacity>
-              <SharesSelection causeAreaOrgs={causeAreaOrgs} />
-              <SharesSum causeArea={causeAreaOrgs.name} />
-            </AnimateHeight>
+            <SharesSelection causeAreaOrgs={causeAreaOrgs} open={shareType == ShareType.CUSTOM} />
           </div>
         );
       })}
