+++ components/shared/components/Widget/components/shared/Toggle/Toggle.style.tsx
@@ -1,22 +1,27 @@
 import styled from "styled-components";
 
-export const StyledToggle = styled.div<{ active: boolean }>`
+export const StyledToggle = styled.button<{ active: boolean }>`
   display: flex;
   align-items: center;
   justify-content: flex-start;
-  width: 48px;
+  width: 44px;
   height: 24px;
   border-radius: 12px;
   background-color: ${({ active }) => (active ? "var(--primary)" : "var(--inactive)")};
   cursor: pointer;
   transition: background-color 0.2s ease-in-out;
   position: relative;
+
+  &:focus {
+    outline: 2px solid var(--primary);
+    outline-offset: 3px;
+  }
 `;
 
 export const StyledToggleCircle = styled.div<{ active: boolean }>`
-  height: 90%;
+  height: 20px;
   aspect-ratio: 1;
-  margin-left: 5%;
+  margin-left: 2px;
   border-radius: 50%;
   background-color: var(--secondary);
   transition: transform 0.2s ease-in-out;
